Improve performance by removing conversion to double for Rtree envelope

+++ src/BlazorDatasheet.DataStructures/Store/RegionDataStore.cs
@@ -531,7 +531,7 @@ public class RegionDataStore<T> : ISparseSource, IRowSource, IStore<T, RegionRes
 
     public int GetNextNonEmptyIndex(int index)
     {
-        var regions = Tree.Search(new Envelope(0, 0, double.MaxValue, double.MaxValue)).Select(x => x.Region);
+        var regions = Tree.Search(new Envelope(0, 0, int.MaxValue, int.MaxValue)).Select(x => x.Region);
         int nextRowIndex = int.MaxValue;
         foreach (var region in regions)
         {
@@ -554,7 +554,7 @@ public class RegionDataStore<T> : ISparseSource, IRowSource, IStore<T, RegionRes
         int dir = Math.Sign(colDir);
 
         var searchEnv = Math.Sign(dir) == 1
-            ? new Envelope(col + 1, row, double.MaxValue, row + 1)
+            ? new Envelope(col + 1, row, int.MaxValue, row + 1)
             : new Envelope(0, row, col, row + 1);
 
         var regions = Tree.Search(searchEnv).Select(x => x.Region);
